move verifier to a different crate
