+++ crates/graph/src/graph.rs
@@ -18,15 +18,13 @@ use luminair_air::{
         sin::table::{SinColumn, SinTraceTable},
         sqrt::table::{SqrtColumn, SqrtTraceTable},
         sum_reduce::table::{SumReduceColumn, SumReduceTraceTable},
-        LuminairComponents, LuminairInteractionElements,
     },
     pie::{
         ExecutionResources, InputInfo, LuminairPie, NodeInfo, OpCounter, OutputInfo, TraceTable,
     },
-    preprocessed::{lookups_to_preprocessed_column, LookupLayout, PreProcessedTrace, Range},
+    preprocessed::{LookupLayout, Range},
     settings::CircuitSettings,
-    utils::{calculate_log_size, log_sum_valid},
-    LuminairProof,
+    utils::calculate_log_size,
 };
 use luminair_utils::LuminairError;
 use luminal::{
@@ -34,15 +32,6 @@ use luminal::{
     prelude::{petgraph::visit::EdgeRef, *},
 };
 use numerair::Fixed;
-use stwo_prover::{
-    constraint_framework::{INTERACTION_TRACE_IDX, ORIGINAL_TRACE_IDX, PREPROCESSED_TRACE_IDX},
-    core::{
-        channel::Blake2sChannel,
-        pcs::{CommitmentSchemeVerifier, PcsConfig},
-        prover::verify,
-        vcs::blake2_merkle::{Blake2sMerkleChannel, Blake2sMerkleHasher},
-    },
-};
 
 /// Trait defining the core functionality of a LuminAIR computation graph.
 ///
@@ -54,13 +43,6 @@ pub trait LuminairGraph {
 
     /// Generates an execution trace for the graph's computation.
     fn gen_trace(&mut self, settings: &mut CircuitSettings) -> Result<LuminairPie, LuminairError>;
-
-    /// Verifies a proof to ensure integrity of graph's computation.
-    fn verify(
-        &self,
-        proof: LuminairProof<Blake2sMerkleHasher>,
-        settings: CircuitSettings,
-    ) -> Result<(), LuminairError>;
 }
 
 /// Implementation of `LuminairGraph` for the `luminal::Graph` struct.
@@ -438,93 +420,6 @@ impl LuminairGraph for Graph {
                 max_log_size,
             },
         })
-    }
-
-    /// Verifies a STWO proof.
-    ///
-    /// Takes a `LuminairProof` and `CircuitSettings` as input.
-    /// It orchestrates the STWO verification protocol:
-    /// 1. Sets up the verifier, channel, and commitment scheme.
-    /// 2. Reads commitments for preprocessed, main, and interaction traces from the proof.
-    /// 3. Derives interaction elements using Fiat-Shamir.
-    /// 4. Constructs the AIR components (constraints) based on the claims and interaction elements.
-    /// 5. Verifies the STARK proof.
-    /// Returns `Ok(())` if the proof is valid, otherwise returns a `LuminairError`.
-    fn verify(
-        &self,
-        LuminairProof {
-            claim,
-            interaction_claim,
-            proof,
-        }: LuminairProof<Blake2sMerkleHasher>,
-        settings: CircuitSettings,
-    ) -> Result<(), LuminairError> {
-        // Convert lookups in circuit settings to preprocessed column.
-        let lut_cols = lookups_to_preprocessed_column(&settings.lookups);
-        let preprocessed_trace = PreProcessedTrace::new(lut_cols);
-
-        // ┌──────────────────────────┐
-        // │     Protocol Setup       │
-        // └──────────────────────────┘
-        let config = PcsConfig::default();
-        let channel = &mut Blake2sChannel::default();
-        let commitment_scheme_verifier =
-            &mut CommitmentSchemeVerifier::<Blake2sMerkleChannel>::new(config);
-
-        // Prepare log sizes for each phase
-        let mut log_sizes = claim.log_sizes();
-        log_sizes[PREPROCESSED_TRACE_IDX] = preprocessed_trace.log_sizes();
-
-        // ┌───────────────────────────────────────────────┐
-        // │   Interaction Phase 0 - Preprocessed Trace    │
-        // └───────────────────────────────────────────────┘
-        commitment_scheme_verifier.commit(
-            proof.commitments[PREPROCESSED_TRACE_IDX],
-            &log_sizes[PREPROCESSED_TRACE_IDX],
-            channel,
-        );
-
-        // ┌───────────────────────────────────────┐
-        // │    Interaction Phase 1 - Main Trace   │
-        // └───────────────────────────────────────┘
-        claim.mix_into(channel);
-        commitment_scheme_verifier.commit(
-            proof.commitments[ORIGINAL_TRACE_IDX],
-            &log_sizes[ORIGINAL_TRACE_IDX],
-            channel,
-        );
-
-        // ┌───────────────────────────────────────────────┐
-        // │    Interaction Phase 2 - Interaction Trace    │
-        // └───────────────────────────────────────────────┘
-        let interaction_elements = LuminairInteractionElements::draw(channel);
-
-        // Validate LogUp sum
-        if !log_sum_valid(&interaction_claim) {
-            return Err(LuminairError::InvalidLogUp("Invalid LogUp sum".to_string()));
-        }
-
-        interaction_claim.mix_into(channel);
-        commitment_scheme_verifier.commit(
-            proof.commitments[INTERACTION_TRACE_IDX],
-            &log_sizes[INTERACTION_TRACE_IDX],
-            channel,
-        );
-
-        // ┌──────────────────────────┐
-        // │    Proof Verification    │
-        // └──────────────────────────┘
-        let component_builder = LuminairComponents::new(
-            &claim,
-            &interaction_elements,
-            &interaction_claim,
-            &preprocessed_trace,
-            &settings.lookups,
-        );
-        let components = component_builder.components();
-
-        verify(&components, channel, commitment_scheme_verifier, proof)
-            .map_err(LuminairError::StwoVerifierError)
     }
 }
