+++ crates/graph/src/tests/expansions.rs
@@ -1,5 +1,6 @@
 use crate::{graph::LuminairGraph, StwoCompiler};
 use luminair_prover::prover::prove;
+use luminair_verifier::verifier::verify;
 use luminal_cpu::CPUCompiler;
 use luminal::prelude::*;
 use rand::{rngs::StdRng, SeedableRng};
@@ -45,7 +46,7 @@ where
         .map_err(|e| format!("Trace generation failed for {}: {:?}", name, e))?;
     let proof = prove(trace, settings.clone())
         .map_err(|e| format!("Proof generation failed for {}: {:?}", name, e))?;
-    cx.verify(proof, settings)
+    verify(proof, settings)
         .map_err(|e| format!("Proof verification failed for {}: {:?}", name, e))?;
 
     let stwo_result = result.data();
