+++ crates/graph/src/tests/mod.rs
@@ -27,7 +27,7 @@ macro_rules! single_unary_test {
                 c.drop();
                 let trace = cx.gen_trace(&mut settings).expect("Trace generation failed");
                 let proof = prove(trace, settings.clone()).expect("Proof generation failed");
-                cx.verify(proof, settings.clone()).expect("Proof verification failed");
+                verify(proof, settings.clone()).expect("Proof verification failed");
                 // Retrieve output data
                 let stwo_output = c.data();
 
@@ -120,7 +120,7 @@ macro_rules! single_binary_test {
                 c.drop();
                 let trace = cx.gen_trace(&mut settings).expect("Trace generation failed");
                 let proof = prove(trace, settings.clone()).expect("Proof generation failed");
-                cx.verify(proof, settings).expect("Proof verification failed");
+                verify(proof, settings).expect("Proof verification failed");
                 // Retrieve output data
                 let stwo_output = c.data();
