+++ crates/verifiers/rust/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "luminair_verifier"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[dependencies]
+luminair_air = { path = "../../air" }
+stwo-prover.workspace = true
+luminair_utils = { path = "../../utils" }
+tracing.workspace = true
